Small cleanup on symbol processor
